ScaleInvariantSNR
